requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pyjwt==2.9.0
cryptography==43.0.1
httpx==0.27.2
python-dateutil==2.9.0.post0
